requirements.txt
blinker==1.9.0
chardet==5.2.0
click==8.1.8
contourpy==1.3.1
cycler==0.12.1
Flask==3.1.0
Flask-SQLAlchemy==3.1.1
fonttools==4.56.0
itsdangerous==2.2.0
Jinja2==3.1.5
kiwisolver==1.4.8
MarkupSafe==3.0.2
matplotlib==3.10.1
natsort==8.4.0
numpy==2.2.3
packaging==24.2
pillow==11.1.0
pyparsing==3.2.1
python-dateutil==2.9.0.post0
reportlab==4.2.5
SimpleITK==2.4.1
six==1.17.0
SQLAlchemy==2.0.37
typing_extensions==4.12.2
Werkzeug==3.1.3
